refactor(checker): split reference_helpers.rs under 2000-line cap
Extract the homomorphic self-referential mapped-type detection cluster
(alias_body_is_direct_homomorphic_self_mapped, template_is_direct_identity_self_ref,
mapped_type_param_has_keyof_alias_param_constraint,
source_is_homomorphic_self_mapped_tuple_arg_vs_tuple_target) from the end of
the impl block into a new sibling child module reference_helpers_homomorphic.rs,
declared via #[path] mod, mirroring the existing child-module pattern. Pure
code-move; behavior identical.

AgentName: StudioOpus
